requirements.txt
Pillow==9.5.0
praw==7.7.0
prawcore==2.3.0